checkstyle, unused imports
git-svn-id: https://versjonskontroll.adeo.no/svn/nav-sbl/dialogarena/common/trunk@16686 6ec884bf-e345-0410-bd42-d282eb737ba0

+++ pdf/src/main/java/no/nav/sbl/dialogarena/pdf/PdfMerger.java
@@ -10,7 +10,6 @@ import org.slf4j.LoggerFactory;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.util.Collection;
 
 public class PdfMerger implements Transformer<Iterable<byte[]>, byte[]> {
 
@@ -36,7 +35,7 @@ public class PdfMerger implements Transformer<Iterable<byte[]>, byte[]> {
         } catch (IOException | DocumentException e) {
             throw new RuntimeException(e);
         } finally {
-            double elapsedTime = (double) (System.currentTimeMillis() - start) / 1000.0;
+            double elapsedTime = (System.currentTimeMillis() - start) / 1000.0;
             logger.debug("Merget {} pdf-dokumenter på {} sekunder", numberOfMergedDocuments, elapsedTime);
             if (merger != null) {
                 merger.close();
